seperate build and test builds
