+++ .github/workflows/check.yml
@@ -25,7 +25,7 @@ jobs:
         run: rustup show
       - uses: Swatinem/rust-cache@v2
       - name: Check lint
-        run: pip install cairo-lang; cargo clippy
+        run: cargo clippy
         
   tests:
     runs-on: ubuntu-latest
@@ -57,4 +57,4 @@ jobs:
       - name: Run udeps
         uses: actions-rs/cargo@v1
         with:
-          command: udeps
+          command: udeps --all-targets
