+++ Cargo.toml
@@ -13,7 +13,6 @@ serde_json = "1.0.105"
 thiserror = "1.0.48"
 zip = { version = "0.6.6", features = ["deflate-zlib"] }
 hex = "0.4.3"
-starknet = { git = "https://github.com/xJonathanLEI/starknet-rs" }
 uuid = { version = "1.4.0", features = ["v4", "serde"] }
 reqwest = { version = "0.11.18", features = ["blocking", "json"] }
 bitvec = { version = "1.0.1", features = ["serde"] }
